pyproject.toml
[project]
name = "mediainfo-bot"
version = "0.1.0"
description = "Python bot for getting infos"
authors = [
    {name = "William Zeni",email = "williamzeni56@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot (>=22.3,<23.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "tmdbsimple (>=2.9.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
